Fix type for iface.fd in startLoop call

+++ app/src/main/java/com/palazik/vpn/service/palazikVpnService.kt
@@ -157,7 +157,7 @@ class palazikVpnService : VpnService() {
 
                 // Pass actual TUN fd — xray reads packets from it via the "tun" inbound.
                 // v2rayNG: tunFd = vpnInterface?.fd ?: 0
-                controller.startLoop(config, iface.fd.toLong())
+                controller.startLoop(config, iface.fd)
 
                 _connectionState.value = ServiceState.RUNNING
                 updateNotification("Connected — ${profile.name}")
